ruthi/requirements.txt
streamlit 
google-generativeai
Pillow
speechrecognition
folium
streamlit-folium
requests 
pyttsx3s